+++ liveOnReboot/liveOn/View/SignIn/GettingStartView.swift
@@ -10,14 +10,13 @@ import Moya
 import AuthenticationServices
 
 struct GettingStartView: View {
-    @ObservedObject var authenticationViewModel: AuthenticationViewModel = AuthenticationViewModel()
+    @ObservedObject var appleSignInViewModel: AppleSignInViewModel = AppleSignInViewModel()
     @State private var authenticationData: String = ""
     @State private var isActive: Bool = false
     
     let authProvider = MoyaProvider<AuthAPI>(plugins: [NetworkLoggerPlugin(verbose: true)])
     
     var body: some View {
-
         VStack {
             VStack(alignment: .leading, spacing: 4) {
                 
@@ -43,9 +42,8 @@ struct GettingStartView: View {
                 }
                 SignInWithAppleButton(.signIn, onRequest: { request in request.requestedScopes = []
                 }, onCompletion: { result in
-                    authenticationData = authenticationViewModel.didFinishAuthentication(result: result)
-                   // print("------\(result)")
-                    isActive.toggle()
+                    authenticationData = appleSignInViewModel.didFinishAppleSignin(result: result)
+//                    isActive.toggle()
                 })
                 .frame(width: 280, height: 60)
                 .padding(.top, 50)
@@ -55,26 +53,10 @@ struct GettingStartView: View {
         .ignoresSafeArea()
         .background(Color.backgroundGray)
     }
-    // didfinishAUthenticaiton에 result를 넣으면, appleUser.identityToken을 산출해주고
-    // appleUser.identityToken을 authNetwrokService의 login 함수에 넣으면 디코딩된 결과를 반환해준다.
-}
-
-private func showAppleLoginView() {
-    
-    let provider = ASAuthorizationAppleIDProvider()
-    
-    let request = provider.createRequest()
-    
-    request.requestedScopes = [.fullName, .email]
-    
-    let controller = ASAuthorizationController(authorizationRequests: [request])
-    
-    controller.performRequests()
-    
 }
 
 struct GettingStart_Previews: PreviewProvider {
     static var previews: some View {
-        GettingStartView(authenticationViewModel: AuthenticationViewModel())
+        GettingStartView(appleSignInViewModel: AppleSignInViewModel())
     }
 }
